user.is_authenticated add

+++ community/views.py
@@ -19,10 +19,13 @@ def reviews_list_create(request,movie_pk):
         serializer = serializers.ReviewListSerializers(reviews, many=True)
         return Response(serializer.data,status=status.HTTP_200_OK)
     else:
-        serializer = serializers.ReviewSerializers(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(user=request.user, movie_id=movie_pk)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user.is_authenticated:
+            serializer = serializers.ReviewSerializers(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(user=request.user, movie_id=movie_pk)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 @api_view(['GET','PUT','DELETE'])
 def review_detail(request,review_pk):
@@ -30,17 +33,21 @@ def review_detail(request,review_pk):
     if request.method == 'GET':
         serializer = serializers.ReviewSerializers(instance=review)
         return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = serializers.ReviewSerializers(instance=review,data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
-    elif request.method == 'DELETE':
-        review.delete()
-        data = {
-            'delete' : f'리뷰 {review_pk} 번이 삭제되었습니다'
-        }
-        return Response(data, status=status.HTTP_204_NO_CONTENT)
+    elif request.user.is_authenticated:
+        if request.method == 'PUT':
+            serializer = serializers.ReviewSerializers(instance=review,data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+        elif request.method == 'DELETE':
+            review.delete()
+            data = {
+                'delete' : f'리뷰 {review_pk} 번이 삭제되었습니다'
+            }
+            return Response(data, status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+
 
 @api_view(['GET','POST'])
 def review_comments_list_create(request,review_pk):
@@ -49,29 +56,38 @@ def review_comments_list_create(request,review_pk):
         serializer = serializers.ReviewCommentListSerializers(comments, many=True)
         return Response(serializer.data,status=status.HTTP_200_OK)
     elif request.method == 'POST':
-        serializer = serializers.ReviewCommentSerializers(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(user=request.user, review_id= review_pk)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user.is_authenticated:
+            serializer = serializers.ReviewCommentSerializers(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(user=request.user, review_id= review_pk)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
 
 @api_view(['GET','PUT','DELETE'])
 def review_comments_detail(request,review_comment_pk):
     comment = get_object_or_404(Review_comment,pk=review_comment_pk)
-    if request.method == 'PUT':
-        serializer = serializers.ReviewCommentSerializers(instance=comment,data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data,status=status.HTTP_200_OK)
-    elif request.method == 'DELETE':
-        comment.delete()
-        data = {
-            'delete': f'댓글{review_comment_pk}번이 삭제되었습니다.'
-        }
-        return Response(data,status=status.HTTP_204_NO_CONTENT)
-    serializer = serializers.ReviewCommentSerializers(comment)
-    return Response(serializer.data)
+    if request.method == 'GET':
+        serializer = serializers.ReviewCommentSerializers(comment)
+        return Response(serializer.data)
+    elif request.user.is_authenticated:
+        if request.method == 'PUT':
+            serializer = serializers.ReviewCommentSerializers(instance=comment,data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data,status=status.HTTP_200_OK)
+        elif request.method == 'DELETE':
+            comment.delete()
+            data = {
+                'delete': f'댓글{review_comment_pk}번이 삭제되었습니다.'
+            }
+            return Response(data,status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+        
 
-
+@api_view(['POST'])
 def review_like(request, review_pk):
     if request.user.is_authenticated:
         review = get_object_or_404(Review,pk=review_pk)
@@ -89,28 +105,33 @@ def chats_list_create(request):
         serializer = serializers.ChatboardListSerializers(chats, many=True)
         return Response(serializer.data,status=status.HTTP_200_OK)
     else:
-        serializer = serializers.ChatboardSerializers(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(user=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user.is_authenticated:
+            serializer = serializers.ChatboardSerializers(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(user=request.user)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 @api_view(['GET','PUT','DELETE'])
 def chat_detail(request, chatboard_pk):
     chat = get_object_or_404(Chatboard,pk=chatboard_pk)
     if request.method == 'GET':
-        serializer = serializers.ReviewSerializers(instance=chat)
+        serializer = serializers.ChatboardSerializers(instance=chat)
         return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = serializers.ReviewSerializers(instance=chat,data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
-    elif request.method == 'DELETE':
-        chat.delete()
-        data = {
-            'delete' : f'게시글 {chatboard_pk} 번이 삭제되었습니다'
-        }
-        return Response(data, status=status.HTTP_204_NO_CONTENT)
+    elif request.user.is_authenticated:
+        if request.method == 'PUT':
+            serializer = serializers.ChatboardSerializers(instance=chat,data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+        elif request.method == 'DELETE':
+            chat.delete()
+            data = {
+                'delete' : f'게시글 {chatboard_pk} 번이 삭제되었습니다'
+            }
+            return Response(data, status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['GET','POST'])
@@ -120,25 +141,32 @@ def chat_comments_list_create(request, chatboard_pk):
         serializer = serializers.ChatboardCommentListSerializers(comments, many=True)
         return Response(serializer.data,status=status.HTTP_200_OK)
     elif request.method == 'POST':
-        serializer = serializers.ChatboardCommentSerializers(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(user=request.user, chatboard_id= chatboard_pk)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user.is_authenticated:
+            serializer = serializers.ChatboardCommentSerializers(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(user=request.user, chatboard_id= chatboard_pk)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['GET','PUT','DELETE'])
 def chat_comments_detail(request,chat_comment_pk):
     comment = get_object_or_404(Chatboard_comment,pk=chat_comment_pk)
-    if request.method == 'PUT':
-        serializer = serializers.ChatboardCommentSerializers(instance=comment,data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data,status=status.HTTP_200_OK)
-    elif request.method == 'DELETE':
-        comment.delete()
-        data = {
-            'delete': f'댓글{chat_comment_pk}번이 삭제되었습니다.'
-        }
-        return Response(data,status=status.HTTP_204_NO_CONTENT)
-    serializer = serializers.ChatboardCommentSerializers(comment)
-    return Response(serializer.data)
+    if request.method == 'GET':
+        serializer = serializers.ChatboardCommentSerializers(comment)
+        return Response(serializer.data)
+    elif request.user.is_authenticated:
+        if request.method == 'PUT':
+            serializer = serializers.ChatboardCommentSerializers(instance=comment,data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data,status=status.HTTP_200_OK)
+        elif request.method == 'DELETE':
+            comment.delete()
+            data = {
+                'delete': f'댓글{chat_comment_pk}번이 삭제되었습니다.'
+            }
+            return Response(data,status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
